+++ packages/antd-management-fast-formily/CHANGELOG.md
@@ -1,5 +1,15 @@
 # antd-management-fast-formily
 
+## 1.0.45
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.45
+  - antd-management-fast-component@2.6.110
+  - antd-management-fast-common@2.6.71
+
 ## 1.0.44
 
 ### Patch Changes