+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.11.128
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.110
+  - antd-management-fast-common@2.6.71
+
 ## 2.11.127
 
 ### Patch Changes